refactor: controller 파일명 컨벤션 적용
